requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-wtf==1.2.2
flask-bcrypt==1.0.1
python-dotenv==1.0.1
email-validator==2.2.0
psycopg2-binary==2.9.10
mailersend==0.6.0
